# Conflicts:
#	segmentation/annotation.py
#	segmentation/filter.py
#	segmentation/test.py
#	ultralytics/data/augment.py
#	ultralytics/data/dataset.py

+++ segmentation/filter.py
@@ -86,8 +86,6 @@ def ann_filter():
             os.remove(os.path.join(he_dir, he))
             os.remove(ann_path)
             logger.info(f'{he}没有腺体')
-
-
 # ann_filter()
 
 # 图像大小过滤
@@ -137,22 +135,24 @@ def size_filter1():
 
 # 指定你的文件夹路径
 def split_data():
-    he_dir = r'/NAS2/Data1/lbliao/Data/MXB/Seg-Relabel/dataset/images_2048'
-    ann_path = r'/NAS2/Data1/lbliao/Data/MXB/Seg-Relabel/dataset/labels_2048/'
+    he_dir = r'/NAS2/Data1/lbliao/Data/MXB/Detection/cellvit+/dataset/images'
+    ann_path = r'/NAS2/Data1/lbliao/Data/MXB/Detection/cellvit+/dataset/labels'
 
-    img_dir = r'/NAS2/Data1/lbliao/Data/MXB/Seg-Relabel/dataset/2048/'
+    img_dir = r'/NAS2/Data1/lbliao/Data/MXB/Detection/cellvit+/dataset/2048'
     os.makedirs(img_dir, exist_ok=True)
 
     he_imgs = os.listdir(he_dir)
     random.shuffle(he_imgs)
     length = len(he_imgs)
-    train = int(length * 0.9)
+    train = int(length * 0.7)
     for i in range(train):
         base, ext = os.path.splitext(he_imgs[i])
         img = os.path.join(img_dir, 'train/images')
         label = os.path.join(img_dir, 'train/labels')
         os.makedirs(img, exist_ok=True)
         os.makedirs(label, exist_ok=True)
+        if os.path.exists(os.path.join(img, he_imgs[i])):
+            continue
         shutil.copy(os.path.join(he_dir, he_imgs[i]), os.path.join(img, he_imgs[i]))
         shutil.copy(os.path.join(ann_path, f'{base}.txt'), os.path.join(label, f'{base}.txt'))
 
@@ -184,6 +184,7 @@ def img_resize():
 
 
 # img_resize()
+
 
 
 # split_data()
@@ -284,24 +285,5 @@ def my_test():
         elif file.endswith('.png'):
             os.rename(file_path, os.path.join(img_dir, file))
 
-
 split_data()
 # my_test()
-
-def my_test2():
-    ann_path = '/NAS2/Data1/lbliao/Data/MXB/Seg-Relabel/dataset/labels_2048'
-    img_path = '/NAS2/Data1/lbliao/Data/MXB/Seg-Relabel/dataset/images_2048'
-    imgs = os.listdir(img_path)
-    for img in imgs:
-        base, _ = os.path.splitext(img)
-        if not os.path.isfile(os.path.join(ann_path, f'{base}.txt')):
-            os.remove(os.path.join(img_path, img))
-            logger.info(f'remove image {img}')
-    anns = os.listdir(ann_path)
-    for ann in anns:
-        base, _ = os.path.splitext(ann)
-        if not os.path.isfile(os.path.join(ann_path, f'{base}.txt')):
-            os.remove(os.path.join(ann_path, ann))
-            logger.info(f'remove ann {ann}')
-
-# my_test2()
